bugfix: wrong number type in AlgoStats1D
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@24383 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ common/src/geogebra/common/kernel/algos/AlgoStats1D.java
@@ -152,7 +152,7 @@ public abstract class AlgoStats1D extends AlgoElement {
 		// Product[{1, 2, 3}, {2, 4, 8}]
 		
 		int truncate;
-		int size = geoList.size();
+		double size = geoList.size();
 
 		if (Truncate != null) {
 			truncate = (int) Truncate.getDouble();
@@ -252,7 +252,7 @@ public abstract class AlgoStats1D extends AlgoElement {
 
 			}
 
-			size = (int) sumFreq;
+			size = sumFreq;
 		}
 
 		mu = sumVal / size;
